refactor: lightweight component
- lightweight chart component는 lightweight 기본 구현에
필요한 것만 구현하고, 심볼 차트를 구현하는 것에만 필요한
것들은 lightweight chart component를 활용한 컴포넌트로 분리하여
의존되지 않도록 수정함(추후 다른 차트가 필요할 경우 로직이
겹치면 분리하기가 힘들기 때문에 최대한 독립적으로 유지되어야 함)

- zIndex 수정

- 차트데이터가 없으면 없다는 메시지만 보여주고, 에러인 경우에만
관련메시지와 함께 다시 로드할 수 있는 UI를 제공
