--- src/components/Button/ButtonContainer.tsx
@@ -1,17 +1,0 @@
-"use client"
-
-import React from 'react'
-import styled from '@emotion/styled'
-
-const Container = styled.div`
-  display:flex;
-  justify-content: space-between;
-  gap:6px;
-  font-size: 24px;
-`
-
-export default function ButtonContainer({children}:{children: React.ReactNode}) {
-  return (
-    <Container>{children}</Container>
-  )
-}
