+++ src/components/Chart/ChartLoding.tsx
@@ -2,11 +2,15 @@ import { LineChartLoading } from '@/components/Chart/LineChartLoading';
 
 function ChartLoading() {
   return (
-    <>
+    <div>
       <div className="w-full">
         <LineChartLoading />
       </div>
-      <div className="pb-7 pt-4">
+      <div className="divide-y-2 divide-dashed divide-[#A7A7A7]">
+        <div className="h-2" />
+        <div className="h-2 pt-0.5" />
+      </div>
+      <div className="pb-7">
         <div className="skeleton_loading mt-4 flex items-end justify-between gap-1">
           <div className="h-5 flex-1 rounded-tl-lg rounded-tr-lg bg-gray-300"></div>
           <div className="h-2 flex-1 rounded-tl-lg rounded-tr-lg bg-gray-300"></div>
@@ -20,7 +24,7 @@ function ChartLoading() {
           <div className="h-20 flex-1 rounded-tl-lg rounded-tr-lg bg-gray-300"></div>
         </div>
       </div>
-    </>
+    </div>
   );
 }
